CW06.10

+++ MVC/src/main/java/ua/controller/WishListAndOrderController.java
@@ -2,23 +2,38 @@ package ua.controller;
 
 import java.security.Principal;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.WebDataBinder;
+import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 
-import ua.entity.MyOrder;
+
+
+
+
+import ua.entity.Carrier;
+import ua.entity.City;
 import ua.form.AddOrderForm;
 import ua.form.DeliveryForm;
 import ua.service.CarrierService;
 import ua.service.CityService;
+import ua.service.DeliveryService;
 import ua.service.OrderService;
 import ua.service.ProducerService;
 import ua.service.ProductService;
 import ua.service.ProductTypeService;
 import ua.service.UserService;
+import ua.service.implementation.editor.CarrierEditor;
+import ua.service.implementation.editor.CityEditor;
+import ua.service.implementation.validator.DeliveryValidator;
 
 @Controller
 public class WishListAndOrderController {
@@ -36,12 +51,23 @@ public class WishListAndOrderController {
 	private UserService userService;
 	@Autowired
 	private OrderService orderService;
+	@Autowired
+	private DeliveryService deliveryService;
 
 	@ModelAttribute("addOrderForm")
 	public AddOrderForm getOrderForm() {
 		return new AddOrderForm();
 	}
-
+	@InitBinder("deliveryForm")
+	protected void initBinderDelivery(WebDataBinder binderDelivery) {
+		binderDelivery.registerCustomEditor(City.class, new CityEditor(
+				cityService));
+		binderDelivery.registerCustomEditor(Carrier.class, new CarrierEditor(
+				carrierService));
+		binderDelivery.setValidator(new DeliveryValidator(deliveryService));
+	}
+	
+	
 	@ModelAttribute("deliveryForm")
 	public DeliveryForm getDeliveryForm() {
 		return new DeliveryForm();
@@ -58,21 +84,37 @@ public class WishListAndOrderController {
 
 	@RequestMapping("/user/order")
 	public String showProductOrder(Model model,
-			@ModelAttribute("addOrderForm") AddOrderForm addForm, @ModelAttribute("deliveryForm") DeliveryForm deliveryForm, Principal principal) {
-		MyOrder order = orderService.createOrder(principal,addForm.getProductIds());
-		
-			
-		
-		model.addAttribute("products",orderService.findAllProducts(order.getId()));
+			@ModelAttribute("addOrderForm") AddOrderForm addForm,
+			Principal principal) {
+		model.addAttribute("products",
+				productService.findAllInited(addForm.getProductIds()));
 		model.addAttribute("productTypes", productTypeService.findAll());
 		model.addAttribute("producers", producerService.findAll());
 		model.addAttribute("cities", cityService.findAll());
 		model.addAttribute("carriers", carrierService.findAll());
+		return "userOrder";
+	}
+
+	@RequestMapping(value = "/user/order/save", method = RequestMethod.POST)
+	public String saveOrder(Model model,
+			@ModelAttribute("addOrderForm") AddOrderForm addForm,
+			@ModelAttribute("deliveryForm") @Valid DeliveryForm deliveryForm,
+			BindingResult br, Principal principal) {
+		if (br.hasErrors()) {
+			model.addAttribute("products",
+					productService.findAllInited(addForm.getProductIds()));
+			model.addAttribute("productTypes", productTypeService.findAll());
+			model.addAttribute("producers", producerService.findAll());
+			model.addAttribute("cities", cityService.findAll());
+			model.addAttribute("carriers", carrierService.findAll());
+			return "userOrder";
+		}
+
+		orderService.saveOrder(deliveryForm, addForm, principal);
 		for (Integer id : addForm.getProductIds()) {
 			userService.deleteFromWishList(principal, id);
 		}
-		
-		return "userOrder";
+		return "redirect:/user";
 	}
 
 	@RequestMapping(value = "/user/wishlist/delete/{id}")
